API

API documentation

Constants

NodeNumbering.element_adjacencies — Constant.

element_adjacencies

The Dict shows some element types and their node adjacencies. For example the element type Quad4 has four nodes/element so that node No. 1 is adjacent to nodes 2 and 4 etc.

source

Functions

NodeNumbering.adjacency_visualization — Method.

adjacency_visualization(RCM_adjacency::Dict{Int, Vector{Int}})

Visualizes adjacency graph Dicts as matrices. Result can be plotted using PyPlot function matshow(). A part of the Reverse Cuthill-McKee Algorithm.

Examples

julia> RCM_adjacency = Dict{Int, Vector{Int}}(
                       7 => [5, 6, 9],
                       9 => [7, 8],
                       4 => [2, 6],
                       2 => [4, 3, 1],
                       3 => [5, 6, 2, 1],
                       8 => [6, 9],
                       5 => [7, 3],
                       6 => [7, 3, 8, 4],
                       1 => [3, 2]);

julia> adjacency_visualization(RCM_adjacency)
9×9 Array{Float64,2}:
 1.0  1.0  1.0  0.0  0.0  0.0  0.0  0.0  0.0
 1.0  1.0  1.0  1.0  0.0  0.0  0.0  0.0  0.0
 1.0  1.0  1.0  0.0  1.0  1.0  0.0  0.0  0.0
 0.0  1.0  0.0  1.0  0.0  1.0  0.0  0.0  0.0
 0.0  0.0  1.0  0.0  1.0  0.0  1.0  0.0  0.0
 0.0  0.0  1.0  1.0  0.0  1.0  1.0  1.0  0.0
 0.0  0.0  0.0  0.0  1.0  1.0  1.0  0.0  1.0
 0.0  0.0  0.0  0.0  0.0  1.0  0.0  1.0  1.0
 0.0  0.0  0.0  0.0  0.0  0.0  1.0  1.0  1.0

source

NodeNumbering.create_RCM_adjacency — Method.

create_RCM_adjacency(adjacency::Dict{Int, Vector{Int}}, finalorder::Dict{Int, Int})

Create an adjacency Dict of the RCM ordered nodes. A part of the Reverse Cuthill-McKee Algorithm.

Examples

julia> adjacency = Dict{Int, Vector{Int}}(
                   1 => [3, 8, 9],
                   2 => [3, 8, 7, 6],
                   3 => [1, 2],
                   4 => [8, 9],
                   5 => [7, 8],
                   6 => [2, 7],
                   7 => [5, 2, 6],
                   8 => [1, 2, 4, 5],
                   9 => [1, 4]);

julia> finalorder =  Dict{Int, Int}(
                     7 => 2,
                     9 => 9,
                     4 => 8,
                     2 => 3,
                     3 => 5,
                     5 => 4,
                     8 => 6,
                     6 => 1,
                     1 => 7);

julia> create_RCM_adjacency(adjacency, finalorder)
Dict{Int64,Array{Int64,1}} with 9 entries:
  7 => [5, 6, 9]
  9 => [7, 8]
  4 => [2, 6]
  2 => [4, 3, 1]
  3 => [5, 6, 2, 1]
  8 => [6, 9]
  5 => [7, 3]
  6 => [7, 3, 8, 4]
  1 => [3, 2]

References

Wikipedia contributors. "Adjacency list". Wikipedia, The Free Encyclopedia. Wikipedia, The Free Encyclopedia, 7 Jun. 2017. Web. 17 Jul. 2017. https://en.wikipedia.org/wiki/Adjacency_list

source

NodeNumbering.renumbering — Method.

renumbering(neworder::Array{Int})

Renumber the RCM ordered nodes. A part of the Reverse Cuthill-McKee Algorithm.

Examples

julia> neworder = Int[6, 7, 2, 5, 3, 8, 1, 4, 9];

julia> renumbering(neworder)
Dict{Int64,Int64} with 9 entries:
  7 => 2
  9 => 9
  4 => 8
  2 => 3
  3 => 5
  5 => 4
  8 => 6
  6 => 1
  1 => 7

References

Wikipedia contributors. "Cuthill–McKee algorithm". Wikipedia, The Free Encyclopedia. Wikipedia, The Free Encyclopedia, 11 Jul. 2017. Web. 20 Jul. 2017. https://en.wikipedia.org/wiki/Cuthill%E2%80%93McKee_algorithm

source

NodeNumbering.node_degrees — Method.

node_degrees(adjacency::Dict{Int, Vector{Int}})

Calculate the degrees for the adjacency graph nodes. Node degree = number of nodes a certain node is adjacent to.

Examples

julia> adjacency = Dict{Int, Vector{Int}}(
                   1 => [3, 8, 9],
                   2 => [3, 8, 7],
                   3 => [1, 2],
                   4 => [8, 9],
                   5 => [7, 8],
                   6 => [2, 7],
                   7 => [5, 2, 6],
                   8 => [1, 2, 4, 5],
                   9 => [1, 4]);

julia> node_degrees(adjacency::Dict{Int, Vector{Int}})
Dict{Int64,Int64} with 9 entries:
  7 => 3
  4 => 2
  9 => 2
  2 => 3
  3 => 2
  5 => 2
  8 => 4
  6 => 2
  1 => 3

References

Wikipedia contributors. "Degree (graph theory)". Wikipedia, The Free Encyclopedia. 24 Nov. 2016. Web. 18 Jul. 2017. https://en.wikipedia.org/wiki/Degree_(graph_theory)

source

NodeNumbering.calculate_bandwidth — Method.

calculate_bandwidth(G::Dict{Int, Vector{Int}})

Calculate bandwidth of graph $G$. The bandwidth of the matrix is number $k$ such that $A_{ij} = 0$ if $|i-j| > k$.

Examples

julia> G = Dict(
           1 => [3, 8, 9],
           2 => [3, 8, 7],
           3 => [1, 2],
           4 => [8, 9],
           5 => [7, 8],
           6 => [2, 7],
           7 => [5, 2, 6],
           8 => [1, 2, 4, 5],
           9 => [1, 4]);

julia> calculate_bandwidth(G)
17

References

Wikipedia contributors. "Band matrix." Wikipedia, The Free Encyclopedia. Wikipedia, The Free Encyclopedia, 28 Apr. 2017. Web. 15 Jul. 2017. https://en.wikipedia.org/wiki/Band_matrix.

source

NodeNumbering.RCM — Method.

RCM(adjacency::Dict{Int, Vector{Int}}, degrees::Dict{Int, Int}, P::Int)

Calculate the Reverse Cuthill-McKee Algorithm for the adjacency graph.

Examples

julia> adjacency = Dict(
                   1 => [3, 8, 9],
                   2 => [3, 8, 7],
                   3 => [1, 2],
                   4 => [8, 9],
                   5 => [7, 8],
                   6 => [2, 7],
                   7 => [5, 2, 6],
                   8 => [1, 2, 4, 5],
                   9 => [1, 4]);

julia> degrees = Dict(
                 1 => 3,
                 2 => 4,
                 3 => 2,
                 4 => 2,
                 5 => 2,
                 6 => 2,
                 7 => 3,
                 8 => 4,
                 9 => 2);

julia> P = 9;

julia> RCM(adjacency, degrees, P)
9-element Array{Int64,1}:
 6
 7
 2
 5
 3
 8
 1
 4
 9

References

Wikipedia contributors. "Cuthill–McKee algorithm". Wikipedia, The Free Encyclopedia. Wikipedia, The Free Encyclopedia, 11 Jul. 2017. Web. 17 Jul. 2017. https://en.wikipedia.org/wiki/Cuthill%E2%80%93McKee_algorithm

source

NodeNumbering.create_adjacency_graph — Method.

create_adjacency_graph(elements::Dict{Int, Vector{Int}}, element_types::Dict{Int, Symbol})

Create a Dict that shows all nodes and their adjacencies.

Examples

julia> elements = Dict(
                  1 => [9, 1, 8, 4],
                  2 => [1, 3, 2, 8],
                  3 => [8, 2, 7, 5],
                  4 => [2, 6, 7]);

julia> element_types = Dict(
                       1 => :Quad4,
                       2 => :Quad4,
                       3 => :Quad4,
                       4 => :Tri3);

julia> create_adjacency_graph(elements, element_types)
Dict{Int64,Array{Int64,1}} with 9 entries:
  7 => [2, 6, 5]
  9 => [1, 4]
  4 => [9, 8]
  2 => [6, 7, 3, 8]
  3 => [1, 2]
  8 => [1, 2, 5, 4]
  5 => [8, 7]
  6 => [2, 7]
  1 => [3, 8, 9]

References

Wikipedia contributors. "Adjacency list". Wikipedia, The Free Encyclopedia. Wikipedia, The Free Encyclopedia, 7 Jun. 2017. Web. 17 Jul. 2017. https://en.wikipedia.org/wiki/Adjacency_list

source

Index

NodeNumbering.element_adjacencies

NodeNumbering.RCM

NodeNumbering.adjacency_visualization

NodeNumbering.calculate_bandwidth

NodeNumbering.create_RCM_adjacency

NodeNumbering.create_adjacency_graph

NodeNumbering.node_degrees

NodeNumbering.renumbering